Feature/reply to comments

+++ server/controllers/posts.js
@@ -6,10 +6,18 @@ const {
   updatePost,
   getAllPosts
 } = require('../services/post-services.js');
+
+const { removeLikesFromPost, removeLikesFromComment } = require('../services/like-services');
+
+const {
+  getCommentsOfPost, removeAllPostComments
+} = require('../services/comment-services');
+
 const serverResponse = require('../utils/serverResponse');
 
 const { uploadFile, deleteFile } = require('../services/cloud-services');
 const formatImage = require('../utils/formatMedia.js');
+const { requesterIsAuthenticatedUser } = require('../utils/auth.js');
 
 // @route  POST '/api/posts'
 // @desc   Submit a posts
@@ -18,12 +26,15 @@ const submitPost = async (req, res) => {
   try {
     const buffer = await formatImage(req.file, 600);
     const mediaUrl = await uploadFile(req.file.originalname, buffer);
+
     const post = {
       ...req.body,
       user: req.user.sub,
       media: mediaUrl
     };
+
     const newPost = await createPost(post);
+
     return serverResponse(res, 200, newPost);
   } catch (e) {
     return serverResponse(res, 500, {
@@ -38,6 +49,11 @@ const submitPost = async (req, res) => {
 const getPosts = async (req, res) => {
   try {
     const posts = await getAllPosts(req.query.limit, req.query.skip);
+
+    if (posts.length === 0) {
+      return serverResponse(res, 404, { message: 'No posts found' });
+    }
+
     return serverResponse(res, 200, posts);
   } catch (e) {
     return serverResponse(res, 500, {
@@ -46,18 +62,12 @@ const getPosts = async (req, res) => {
   }
 };
 
-// @route  GET '/api/posts/:userId'
+// @route  GET '/api/posts/:userInfo'
 // @desc   Get posts of a user
 // @access private
 const getPostsOfAUser = async (req, res) => {
   try {
-    const posts = await getAllPostsOfUser(req.params.userId);
-
-    if (!posts.length) {
-      return serverResponse(res, 404, {
-        message: 'there are no posts with requested user id'
-      });
-    }
+    const posts = await getAllPostsOfUser(req.params.userInfo);
 
     return serverResponse(res, 200, posts);
   } catch (e) {
@@ -73,13 +83,14 @@ const getPostsOfAUser = async (req, res) => {
 const getOnePost = async (req, res) => {
   try {
     const post = await getPost(req.params.postId);
+
     if (!post) {
-      return res
-        .status(404)
-        .json({ message: 'no post with requested id' })
-        .end();
+      return serverResponse(res, 404, {
+        message: "Post doesn't exist"
+      });
     }
-    res.status(200).json(post).end();
+
+    return serverResponse(res, 200, post);
   } catch (e) {
     res
       .status(500)
@@ -94,9 +105,29 @@ const getOnePost = async (req, res) => {
 const deletePost = async (req, res) => {
   try {
     const post = await getPost(req.params.postId);
-    await deleteFile(post.media);
-    await removePost(req.params.postId);
-    return serverResponse(res, 200, { message: 'File successfully deleted' });
+
+    if (!post) {
+      return serverResponse(res, 404, {
+        message: "Post doesn't exist"
+      });
+    }
+
+    if (!requesterIsAuthenticatedUser(req.user.sub, post.user)) {
+      return serverResponse(res, 400, {
+        message: 'Unauthorized!'
+      });
+    }
+
+    const comments = await getCommentsOfPost(req.params.postId);
+
+    const removeLikesFromCommentPromises = comments.map(async comment => removeLikesFromComment(comment._id));
+
+    await Promise.all([removeLikesFromPost(req.params.postId),
+      removeAllPostComments(req.params.postId),
+      deleteFile(post.media),
+      removePost(req.params.postId),
+      ...removeLikesFromCommentPromises]);
+    return serverResponse(res, 200, { message: 'Post successfully deleted' });
   } catch (e) {
     return serverResponse(res, 500, {
       message: 'internal error while trying to delete a post'
@@ -109,7 +140,22 @@ const deletePost = async (req, res) => {
 // @access private
 const editPost = async (req, res) => {
   try {
+    const postToEdit = await getPost(req.params.postId);
+
+    if (!postToEdit) {
+      return serverResponse(res, 404, {
+        message: "Post doesn't exist"
+      });
+    }
+
+    if (!requesterIsAuthenticatedUser(req.user.sub, postToEdit.user)) {
+      return serverResponse(res, 400, {
+        message: 'Unauthorized!'
+      });
+    }
+
     const post = await updatePost(req.params.postId, req.body);
+
     return serverResponse(res, 200, post);
   } catch (e) {
     return serverResponse(res, 500, {
